+++ lib/sitemap_generator/adapters/aws_sdk_adapter.rb
@@ -18,15 +18,19 @@ module SitemapGenerator
     # Options:
     #   **Deprecated, use :endpoint instead** :aws_endpoint [String] The object storage endpoint,
     #                                                                if not AWS, e.g. 'https://sfo2.digitaloceanspaces.com'
-    #   **Deprecated, use :access_key_id instead** :access_key_id [String] Your AWS access key id
+    #   **Deprecated, use :access_key_id instead** :aws_access_key_id [String] Your AWS access key id
     #   **Deprecated, use :secret_access_key instead** :aws_secret_access_key [String] Your AWS secret access key
     #   **Deprecated, use :region instead** :aws_region [String] Your AWS region
+    #   :acl [String] The ACL to apply to the uploaded files.  Defaults to 'public-read'.
+    #   :cache_control [String] The cache control headder to apply to the uploaded files.  Defaults to 'private, max-age=0, no-cache'.
     #
     #   All other options you provide are passed directly to the AWS client.
     #   See https://docs.aws.amazon.com/sdk-for-ruby/v2/api/Aws/S3/Client.html#initialize-instance_method
     #   for a full list of supported options.
-    def initialize(bucket, aws_access_key_id: nil, aws_secret_access_key: nil, aws_region: nil, aws_endpoint: nil, **options)
+    def initialize(bucket, aws_access_key_id: nil, aws_secret_access_key: nil, aws_region: nil, aws_endpoint: nil, acl: 'public-read', cache_control: 'private, max-age=0, no-cache', **options)
       @bucket = bucket
+      @acl = acl
+      @cache_control = cache_control
       @options = options
       set_option_unless_set(:access_key_id, aws_access_key_id)
       set_option_unless_set(:secret_access_key, aws_secret_access_key)
@@ -39,11 +43,11 @@ module SitemapGenerator
     def write(location, raw_data)
       SitemapGenerator::FileAdapter.new.write(location, raw_data)
       s3_object = s3_resource.bucket(@bucket).object(location.path_in_public)
-      s3_object.upload_file(location.path,
-        acl: 'public-read',
-        cache_control: 'private, max-age=0, no-cache',
+      s3_object.upload_file(location.path, {
+        acl: @acl,
+        cache_control: @cache_control,
         content_type: location[:compress] ? 'application/x-gzip' : 'application/xml'
-      )
+      }.compact)
     end
 
     private